requirements.txt
a2wsgi==1.10.10
aiosmtplib==4.0.2
aiosqlite==0.21.0
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
apache-airflow==3.1.0
apache-airflow-core==3.1.0
apache-airflow-providers-common-compat==1.7.4
apache-airflow-providers-common-io==1.6.3
apache-airflow-providers-common-sql==1.28.1
apache-airflow-providers-smtp==2.3.1
apache-airflow-providers-standard==1.9.0
apache-airflow-task-sdk==1.1.0
argcomplete==3.6.2
asgiref==3.10.0
async-timeout==5.0.1
attrs==25.4.0
backports.strenum==1.3.1
blinker==1.9.0
blis==1.3.0
cadwyn==5.6.0
catalogue==2.0.10
certifi==2025.10.5
cloudpathlib==0.23.0
colorlog==6.9.0
confection==0.1.5
cron_descriptor==2.0.6
croniter==6.0.0
cymem==2.0.11
dash==3.2.0
Deprecated==1.2.18
dill==0.4.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.119.0
fastapi-cli==0.0.13
Flask==3.1.2
fsspec==2025.9.0
googleapis-common-protos==1.70.0
greenback==1.2.1
greenlet==3.2.4
grpcio==1.75.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
langcodes==3.5.0
language_data==1.3.0
lazy-object-proxy==1.12.0
libcst==1.8.5
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.8.0
msgspec==0.19.0
murmurhash==1.0.13
narwhals==2.8.0
natsort==8.4.0
nest-asyncio==1.6.0
numpy==2.2.6
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
outcome==1.3.0.post0
pandas==2.3.3
pathlib_abc==0.5.1
pathspec==0.12.1
pendulum==3.1.0
plotly==6.3.1
preshed==3.0.10
protobuf==6.32.1
psutil==7.1.0
psycopg2-binary==2.9.11
pydantic==2.11.10  # pinned for Airflow 3.1.0 API compatibility
pydantic_core==2.33.2  # matched to pydantic 2.11.x
pygtrie==2.5.0
PyJWT==2.10.1
PyMuPDF==1.26.5
pytesseract==0.3.13
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
referencing==0.37.0
regex==2025.9.18
requests==2.32.5
retryhttp==1.3.3
retrying==1.4.2
rich==14.2.0
rich-argparse==1.7.1
rich-toolkit==0.15.1
rpds-py==0.27.1
schedule==1.2.2
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.44
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.0
sqlparse==0.5.3
srsly==2.5.1
starlette==0.48.0
structlog==25.4.0
svcs==25.1.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
thinc==8.3.6
tqdm==4.67.1
typer==0.19.2
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.3.3
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.37.0
uvloop==0.21.0
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wirerope==1.0.0
wrapt==1.17.3
zipp==3.23.0
